refactor(dronecan): extract NodeStatus tracking module
Split the node presence/heartbeat concern out of dronecan.c into
dronecan_node_status.c/.h: the live node table (activeNodeCount/
nodeTable[], consumed by dronecan_dna_server.c's isNodeAvailable() via
dronecanGetNodeCount()/dronecanGetNode() to avoid handing out node IDs
already seen on the bus), the incoming NodeStatus broadcast handler
(renamed handle_NodeStatus -> dronecanNodeStatusHandleBroadcast for
external linkage), and our own 1Hz NodeStatus heartbeat send.
dronecan.c's handle_GetNodeInfo() now calls the new
dronecanGetOwnNodeStatus() accessor instead of touching the node_status
struct directly, and process1HzTasks() delegates to
dronecanNodeStatusUpdate() after its own stale-transfer cleanup.

Also drops a dead duplicate stale-node pruning loop: process1HzTasks()
had two back-to-back loops removing nodes past their last-seen
timeout, one using DRONECAN_NODE_STALE_TIMEOUT_MS (10000) and one
hardcoding the same value - the second could never find anything left
to prune since the first already removed it.

Cherry-picked from feature/dronecan-actuator-control (original commit
34aefbc00) directly onto feature/dronecan-dna-server's existing tip:
this is general dronecan.c restructuring in NodeStatus/node-table
territory - driven by dna_server.c's own need for the
dronecanGetNodeCount()/dronecanGetNode() accessors this introduces -
not actuator-control-specific, so it belongs here rather than riding
along with unrelated actuator-output work. Applied as a plain cherry-
pick on top of dna-server's unmodified history (not a full rebase of
its 34 commits onto the updated param-getset branch) to avoid
replaying real pre-existing history irregularities found partway
through an initial rebase attempt.

Note: the include-alphabetization part of the original commit's title
wasn't carried over - dna-server's current header block already has a
different structure (several includes above the USE_DRONECAN guard,
predating this change) than what the original diff assumed, so only
the new dronecan_node_status.h include was added without reordering
the rest.

Full unit test suite (dronecan_application_unittest 29/29,
dronecan_dna_server_unittest 16/16, no failures suite-wide) passes.
SITL builds clean with -Werror.
